refactor(webui): move attachment policy out of channel
